Second commit

+++ app/src/main/java/com/example/taskvmg4/ui/screen/TaskDetailScreen.kt
@@ -134,7 +134,7 @@ fun TaskDetailScreen(
                        onClick = {
                             viewModel.addTask(
                                 Task(
-                                    id = viewModel.id.toInt(),
+                                    id = viewModel.id,
                                     title = viewModel.title,
                                     completed = viewModel.completed,
                                     description = ""
